Implement new contributor detection feature
Added functionality to fetch both all and release-specific Github contributors. This enables automatic detection of new contributors for each release. Elements of this feature were adapted into main.ts and formatter.ts to accommodate this change. Moreover, several unit tests were added to ensure the effectiveness of the feature. Lastly, action.yml was updated to include this feature as an option.